Eliminated tons of warnings when building with gcc 4.8 (default gcc deployed in ubuntu 14.04)
git-svn-id: svn://svn.sphinxsearch.com/sphinx/trunk@4697 406a0c4d-033a-0410-8de8-e80135713968

+++ src/tests.cpp
@@ -3345,7 +3345,7 @@ void TestGeodist()
 			{
 				for ( int b=0; b<360; b+=3, n++ )
 				{
-					double t[4] = { lat, lon, 0, 0 };
+					double t[4] = { double(lat), double(lon), 0, 0 };
 					DestVincenty ( t[0], t[1], b, dist, t+2, t+3 );
 					for ( int j=0; j<4; j++ )
 						dBench.Add ( t[j] );
